oops. forgot that i needed to check before deleting.

+++ build/build.sh
@@ -108,7 +108,13 @@ main() {
     TOOLS_DIR="$(pwd)/tools"
     mkdir -p $TOOLS_DIR
     curl https://codeload.github.com/jason-c-daniels/git-ver/zip/v0.0.1-pre > git-ver.zip
-    rm tools/git-ver*
+    if [[ -f git-ver.zip ]]; then
+        rm git-ver.zip
+    fi
+    if [[ -f tools/git-ver ]]; then
+        rm tools/git-ver*
+    fi
+    
     unzip -jo git-ver.zip **/git-ver* -d ./tools
     export PATH="$TOOLS_DIR":$PATH
